Initial Commit of Rust CNI
This commit brings a Rust implementation of a CNI.
It provides a /32 interface as a gateway in the hostns.
A veth pair with a /32 interface in the container netns.
Sets up Linux Routing and the necessary rules for forwarding.
And optionally enables IP Masquerading (NAT) if the subnet is not
reachable outside the host machine.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>